requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
llama-index==0.10.32
ollama==0.1.7
python-dotenv==1.0.1
python-multipart==0.0.9
